+++ sdk/core/azure-core-http-jdk-httpclient/src/main/java/com/azure/core/http/jdk/httpclient/implementation/JdkHttpResponseSync.java
@@ -145,7 +145,7 @@ public final class JdkHttpResponseSync extends JdkHttpResponseBase {
 
     private BinaryData getBinaryData() {
         if (binaryData == null) {
-            binaryData = BinaryData.fromStream(bodyStream);
+            binaryData = super.getBodyAsBinaryData();
         }
         return binaryData;
     }