+++ sdk/core/azure-core-http-okhttp/src/main/java/com/azure/core/http/okhttp/implementation/OkHttpAsyncResponse.java
@@ -27,6 +27,7 @@ public final class OkHttpAsyncResponse extends OkHttpAsyncResponseBase {
     private static final int BYTE_BUFFER_CHUNK_SIZE = 8192;
 
     private final ResponseBody responseBody;
+    private BinaryData binaryData;
 
     /**
      * Creates an OkHttpAsyncResponse.
@@ -47,7 +48,10 @@ public final class OkHttpAsyncResponse extends OkHttpAsyncResponseBase {
 
     @Override
     public BinaryData getBodyAsBinaryData() {
-        return BinaryData.fromStream(this.responseBody.byteStream());
+        if (binaryData == null) {
+            binaryData = super.getBodyAsBinaryData();
+        }
+        return binaryData;
     }
 
     @Override