+++ sdk/core/azure-core-http-jdk-httpclient/CHANGELOG.md
@@ -8,6 +8,9 @@
 
 ### Bugs Fixed
 
+- Fixed synchronous streaming `BinaryData` response bodies to release JDK HttpClient response resources when reactive
+  consumption is cancelled.
+
 ### Other Changes
 
 ## 1.1.6 (2026-08-12)
@@ -290,5 +293,4 @@ This package's
 [documentation](https://github.com/Azure/azure-sdk-for-java/blob/main/sdk/core/azure-core-http-jdk-httpclient/README.md)
 
 Initial release. Please see the README and wiki for information on the new design.
-